Make schema generator unit-testable
