+++ src/geophires_x_schema_generator/main.py
@@ -2,131 +2,7 @@ import argparse
 import json
 from pathlib import Path
 
-from geophires_x.Model import Model
-
-
-def generate_json_schema() -> dict:
-    input_params_json, output_params_json = Model(enable_geophires_logging_config=False).get_parameters_json()
-    input_params = json.loads(input_params_json)
-
-    properties = {}
-    required = []
-
-    for param_name in input_params:
-        param = input_params[param_name]
-
-        units_val = param['CurrentUnits'] if isinstance(param['CurrentUnits'], str) else None
-        properties[param_name] = {
-            'description': param['ToolTipText'],
-            'type': param['json_parameter_type'],
-            'units': units_val,
-            'category': param['parameter_category'],
-        }
-
-        if param['Required']:
-            required.append(param_name)
-
-    schema = {
-        'definitions': {},
-        '$schema': 'http://json-schema.org/draft-04/schema#',
-        'type': 'object',
-        'title': 'GEOPHIRES Schema',
-        'required': required,
-        'properties': properties,
-    }
-
-    return schema
-
-
-def generate_parameters_reference_rst() -> str:
-    input_params_json, output_params_json = Model(enable_geophires_logging_config=False).get_parameters_json()
-    input_params = json.loads(input_params_json)
-
-    input_rst = """
-    .. list-table:: Input Parameters
-       :header-rows: 1
-
-       * - Name
-         - Description
-         - Preferred Units
-         - Default Value Type
-         - Default Value
-         - Min
-         - Max"""
-
-    for param_name in input_params:
-        param = input_params[param_name]
-
-        # if param['Required']:
-        #     TODO designate required params
-
-        def get_key(k):
-            if k in param and str(param[k]) != '':  # noqa
-                return param[k]  # noqa
-            else:
-                return ''
-
-        min_val = get_key('Min')
-        max_val = get_key('Max')
-
-        if 'AllowableRange' in param:
-            # TODO warn if min/max are defined and at odds with allowable range
-            min_val = min(param['AllowableRange'])
-            max_val = max(param['AllowableRange'])
-
-            # TODO include full AllowableRange in reference (or fully describe if possible using min/max/increment)
-
-        input_rst += f"""\n       * - {param['Name']}
-         - {get_key('ToolTipText')}
-         - {get_key('PreferredUnits')}
-         - {get_key('json_parameter_type')}
-         - {get_key('DefaultValue')}
-         - {min_val}
-         - {max_val}"""
-
-    output_rst = get_output_params_table_rst(output_params_json)
-
-    rst = f"""Parameters
-==========
-
-Input Parameters
-################
-{input_rst}
-
-Output Parameters
-#################
-{output_rst}
-    """
-
-    return rst
-
-
-def get_output_params_table_rst(output_params_json):
-    output_params = json.loads(output_params_json)
-
-    output_rst = """
-    .. list-table:: Output Parameters
-       :header-rows: 1
-
-       * - Name
-         - Preferred Units
-         - Default Value Type"""
-
-    for param_name in output_params:
-        param = output_params[param_name]
-
-        def get_key(k):
-            if k in param and str(param[k]) != '':  # noqa
-                return param[k]  # noqa
-            else:
-                return ''
-
-        output_rst += f"""\n       * - {param['Name']}
-         - {get_key('PreferredUnits')}
-         - {get_key('json_parameter_type')}"""
-
-    return output_rst
-
+from geophires_x_schema_generator import GeophiresXSchemaGenerator
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser()
@@ -146,13 +22,15 @@ if __name__ == '__main__':
 
     build_path = Path(build_dir, 'geophires-request.json')
 
-    schema_json = generate_json_schema()
+    generator = GeophiresXSchemaGenerator()
+
+    schema_json = generator.generate_json_schema()
 
     with open(build_path, 'w') as f:
         f.write(json.dumps(schema_json, indent=2))
         print(f'Wrote schema file to {build_path}.')
 
-    rst = generate_parameters_reference_rst()
+    rst = generator.generate_parameters_reference_rst()
 
     build_path_rst = Path(build_dir, 'parameters.rst')
     with open(build_path_rst, 'w') as f:
