+++ tests/test_geophires_x_schema_generator.py
@@ -1,0 +1,15 @@
+import unittest
+
+from geophires_x_schema_generator import GeophiresXSchemaGenerator
+from tests.base_test_case import BaseTestCase
+
+
+class GeophiresXSchemaGeneratorTestCase(BaseTestCase):
+    def test_parameters_rst(self):
+        g = GeophiresXSchemaGenerator()
+        rst = g.generate_parameters_reference_rst()
+        self.assertIsNotNone(rst)
+
+
+if __name__ == '__main__':
+    unittest.main()
